split staged-tx tables & table impls into separate modules/files
