+++ services/system/StagedTx/service/src/policy.rs
@@ -1,0 +1,45 @@
+use psibase::services::accounts::Wrapper as Accounts;
+use psibase::services::transact::auth_interface::auth_action_structs;
+use psibase::{AccountNumber, Caller, MethodNumber, ServiceCaller};
+
+use crate::service::Wrapper;
+
+pub fn get_auth_service(sender: AccountNumber) -> Option<AccountNumber> {
+    Accounts::call().getAccount(sender).map(|a| a.authService)
+}
+
+pub struct StagedTxPolicy {
+    user: AccountNumber,
+    service_caller: ServiceCaller,
+}
+impl StagedTxPolicy {
+    pub fn new(user: AccountNumber) -> Self {
+        StagedTxPolicy {
+            user,
+            service_caller: ServiceCaller {
+                sender: Wrapper::SERVICE,
+                service: get_auth_service(user).unwrap(),
+            },
+        }
+    }
+
+    pub fn does_auth(&self, accepters: Vec<AccountNumber>) -> bool {
+        self.service_caller.call(
+            MethodNumber::from(auth_action_structs::isAuthSys::ACTION_NAME),
+            auth_action_structs::isAuthSys {
+                sender: self.user,
+                authorizers: accepters,
+            },
+        )
+    }
+
+    pub fn does_reject(&self, rejecters: Vec<AccountNumber>) -> bool {
+        self.service_caller.call(
+            MethodNumber::from(auth_action_structs::isRejectSys::ACTION_NAME),
+            auth_action_structs::isRejectSys {
+                sender: self.user,
+                rejecters,
+            },
+        )
+    }
+}
